+++ packages/core/src/__tests__/session.test.ts
@@ -4,7 +4,7 @@ import {
   getDeviceId,
   getLastRunEndState,
   removeSessionProperty,
-} from "../utils/session";
+} from "../api/session";
 
 const mockAddSessionProperty = jest.fn();
 const mockRemoveSessionProperty = jest.fn();
