+++ packages/core/src/api/component.ts
