+++ packages/core/src/api/error.ts
